feat(tasks): create files for main gulp tasks
