+++ tasks/cache-config.js
@@ -1,0 +1,52 @@
+'use strict';
+
+// Generate config data for the <sw-precache-cache> element.
+// This include a list of files that should be precached, as well as a (hopefully unique) cache
+// id that ensure that multiple PSK projects don't share the same Cache Storage.
+// This task does not run by default, but if you are interested in using service worker caching
+// in your project, please enable it within the 'default' task.
+// See https://github.com/PolymerElements/polymer-starter-kit#enable-service-worker-support
+// for more context.
+
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var glob = require('glob-all');
+var path = require('path');
+var fs = require('fs');
+var crypto = require('crypto');
+var packageJson = require('../package.json');
+
+var DIST = 'dist';
+var dist = function(subpath) {
+  return !subpath ? DIST : path.join(DIST, subpath);
+};
+
+
+gulp.task('cache-config', function(callback) {
+  var dir = dist();
+  var config = {
+    cacheId: packageJson.name || path.basename(__dirname),
+    disabled: false
+  };
+
+  glob([
+    'index.html',
+    './',
+    'bower_components/webcomponentsjs/webcomponents-lite.min.js',
+    '{elements,scripts,styles}/**/*.*'],
+    {cwd: dir}, function(error, files) {
+    if (error) {
+      callback(error);
+    } else {
+      config.precache = files;
+
+      var md5 = crypto.createHash('md5');
+      md5.update(JSON.stringify(config.precache));
+      config.precacheFingerprint = md5.digest('hex');
+
+      var configPath = path.join(dir, 'cache-config.json');
+      fs.writeFile(configPath, JSON.stringify(config), callback);
+    }
+  });
+});
+
