+++ tasks/copy.js
@@ -1,0 +1,35 @@
+'use strict';
+
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var merge = require('merge-stream');
+var path = require('path');
+var DIST = 'dist';
+var dist = function(subpath) {
+  return !subpath ? DIST : path.join(DIST, subpath);
+};
+
+// Copy all files at the root level (app)
+gulp.task('copy', function() {
+  var app = gulp.src([
+    'app/*',
+    '!app/test',
+    '!app/elements',
+    '!app/bower_components',
+    '!app/cache-config.json',
+    '!**/.DS_Store'
+  ], {
+    dot: true
+  }).pipe(gulp.dest(dist()));
+
+  // Copy over only the bower_components we need
+  // These are things which cannot be vulcanized
+  var bower = gulp.src([
+    'app/bower_components/{webcomponentsjs,platinum-sw,sw-toolbox,promise-polyfill}/**/*'
+  ]).pipe(gulp.dest(dist('bower_components')));
+
+  return merge(app, bower)
+    .pipe($.size({
+      title: 'copy'
+    }));
+});
