+++ tasks/serve.js
@@ -1,0 +1,73 @@
+'use strict';
+
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var runSequence = require('run-sequence');
+var browserSync = require('browser-sync');
+var historyApiFallback = require('connect-history-api-fallback');
+var DIST = 'dist';
+var path = require('path');
+var dist = function(subpath) {
+  return !subpath ? DIST : path.join(DIST, subpath);
+};
+
+// Build production files, the default task
+gulp.task('default', ['clean'], function(cb) {
+  // Uncomment 'cache-config' if you are going to use service workers.
+  runSequence(
+    ['ensureFiles', 'copy', 'styles'],
+    'elements',
+    ['images', 'fonts', 'html'],
+    'vulcanize', // 'cache-config',
+    'watch',
+    cb);
+});
+
+// Watch files for changes & reload
+gulp.task('serve', ['styles', 'elements', 'watch'], function() {
+  browserSync({
+    port: 5000,
+    notify: false,
+    logPrefix: 'PSK',
+    snippetOptions: {
+      rule: {
+        match: '<span id="browser-sync-binding"></span>',
+        fn: function(snippet) {
+          return snippet;
+        }
+      }
+    },
+    // Run as an https by uncommenting 'https: true'
+    // Note: this uses an unsigned certificate which on first access
+    //       will present a certificate warning in the browser.
+    // https: true,
+    server: {
+      baseDir: ['.tmp', 'app'],
+      middleware: [historyApiFallback()]
+    }
+  });
+
+});
+
+// Build and serve the output from the dist build
+gulp.task('serve:dist', ['default'], function() {
+  browserSync({
+    port: 5001,
+    notify: false,
+    logPrefix: 'PSK',
+    snippetOptions: {
+      rule: {
+        match: '<span id="browser-sync-binding"></span>',
+        fn: function(snippet) {
+          return snippet;
+        }
+      }
+    },
+    // Run as an https by uncommenting 'https: true'
+    // Note: this uses an unsigned certificate which on first access
+    //       will present a certificate warning in the browser.
+    // https: true,
+    server: dist(),
+    middleware: [historyApiFallback()]
+  });
+});
