+++ tasks/styles.js
@@ -1,0 +1,51 @@
+'use strict';
+
+/*
+* Compile and Automatically Prefix Stylesheets
+*/
+
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var path = require('path');
+var browserSync = require('browser-sync');
+var reload = browserSync.reload;
+var DIST = 'dist';
+var dist = function(subpath) {
+  return !subpath ? DIST : path.join(DIST, subpath);
+};
+
+
+var styleTask = function(stylesPath, srcs) {
+  return gulp.src(srcs.map(function(src) {
+      return path.join('app', stylesPath, src);
+    }))
+    .pipe($.changed(stylesPath, {extension: '.css'}))
+    .pipe($.autoprefixer(AUTOPREFIXER_BROWSERS))
+    .pipe(gulp.dest('.tmp/' + stylesPath))
+    .pipe($.minifyCss())
+    .pipe(gulp.dest(dist(stylesPath)))
+    .pipe($.size({title: stylesPath}));
+};
+
+
+var AUTOPREFIXER_BROWSERS = [
+  'ie >= 10',
+  'ie_mob >= 10',
+  'ff >= 30',
+  'chrome >= 34',
+  'safari >= 7',
+  'opera >= 23',
+  'ios >= 7',
+  'android >= 4.4',
+  'bb >= 10'
+];
+
+// Compile and automatically prefix stylesheets
+gulp.task('styles', function() {
+  return styleTask('styles', ['**/*.css']);
+});
+
+gulp.task('elements', function() {
+  return styleTask('elements', ['**/*.css']);
+});
+
