+++ tasks/vulcanize.js
@@ -1,0 +1,21 @@
+'use strict';
+
+var gulp = require('gulp');
+var $ = require('gulp-load-plugins')();
+var path = require('path');
+var DIST = 'dist';
+var dist = function(subpath) {
+  return !subpath ? DIST : path.join(DIST, subpath);
+};
+
+// Vulcanize granular configuration
+gulp.task('vulcanize', function() {
+  return gulp.src('app/elements/elements.html')
+    .pipe($.vulcanize({
+      stripComments: true,
+      inlineCss: true,
+      inlineScripts: true
+    }))
+    .pipe(gulp.dest(dist('elements')))
+    .pipe($.size({title: 'vulcanize'}));
+});
