+++ tasks/watch.js
@@ -1,0 +1,14 @@
+'use strict';
+var gulp = require('gulp');
+var browserSync = require('browser-sync');
+var reload = browserSync.reload;
+
+// Watch Files For Changes
+gulp.task('watch', function () {
+  gulp.watch(['app/**/*.html', '!app/bower_components/**/*.html'], reload);
+  gulp.watch(['app/styles/**/*.css'], ['styles', reload]);
+  gulp.watch(['app/elements/**/*.css'], ['elements', reload]);
+  gulp.watch(['app/scripts/**/*.js'], reload);
+  gulp.watch(['app/images/**/*'], reload);
+});
+
